Create Readme.md
Introduction
This project is a YouTube clone implemented using HTML, CSS, and JavaScript. The goal of this project is to recreate the basic functionality and design of the popular video-sharing platform YouTube. This clone will allow users to view videos, search for videos, play videos, and interact with the user interface.

Features
Browse and watch videos

Play videos in a video player

Display video information (title, description, uploader, views, likes, etc.)